Begin removing deprecated autoconf 2.13 backward compatibility code. Add AC_CHECK_TYPES macros for uin32/64 types, use uint64_t instead of long long in hcache code.

+++ imap/imap_private.h
@@ -180,7 +180,7 @@ typedef struct
   IMAP_CACHE cache[IMAP_CACHE_LEN];
   unsigned int noclose : 1;
 #ifdef USE_HCACHE
-  unsigned long long uid_validity;
+  uint64_t uid_validity;
 #endif
   
   /* all folder flags - system flags AND keywords */
